requirements.txt
pandas
numpy
scikit-learn
requests
tensorflow-cpu
